backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
joblib
python-multipart
python-jose[cryptography]
bcrypt
sqlalchemy
asyncpg
pydantic[email]
